# apps/ssi-abstraction/swagger.json
{
...
  "info": {
    "title": "ssi-abstraction-agent",
    "version": "0.8.0",
    "description": "Rest endpoint wrapper for using your agent over HTTP"
  },
  "openapi": "3.0.0",
  "paths": {
...
    "/connections/{connectionId}": {
      "get": {
        "operationId": "ConnectionController.getConnectionById",
        "parameters": [
          {
            "in": "path",
            "name": "connectionId",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {}
            },
            "description": "Successful response"
          }
        },
        "summary": "Get connection by id",
        "tags": [
          "Connection"
        ]
      },
      "delete": {
        "operationId": "ConnectionController.deleteConnection",
        "parameters": [
          {
            "in": "path",
            "name": "connectionId",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {}
            },
            "description": "Successful response"
          }
        },
        "summary": "Delete connection",
        "tags": [
          "Connection"
        ]
      }
    },
...
  }
}